+++ cmd.go
@@ -1,4 +1,4 @@
-package command
+package zenith
 
 const (
 	SetCMD  = "SET"
