+++ error.go
@@ -1,4 +1,4 @@
-package errors
+package zenith
 
 import "fmt"
